demo/requirements.txt
django=3.2
django-crispy-forms
django-imagekit